requirements.txt
# Web Framework
Flask==2.3.3
Werkzeug==2.3.7

# Database
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.23

# Environment Variables
python-dotenv==1.0.0

# HTTP Requests (for Paystack API)
requests==2.31.0
urllib3==2.0.7

# Production Server
gunicorn==21.2.0

# PostgreSQL Support (for Render deployment)
psycopg2-binary==2.9.7

# Additional Flask Extensions
Flask-Migrate==4.0.5

# Security & Utilities
MarkupSafe==2.1.3
Jinja2==3.1.2
click==8.1.7
itsdangerous==2.1.2

# Date/Time utilities
pytz==2023.3

# Optional: For better error handling
blinker==1.6.3